refactor(aitools): migrated script tools into script_new and script_review
