+++ Solution.props
@@ -1,5 +1,5 @@
 <Project>
   <PropertyGroup>
-    <SolutionVersion>0.3.2-dev.250525</SolutionVersion>
+    <SolutionVersion>0.3.2-dev.250526</SolutionVersion>
   </PropertyGroup>
 </Project>
